Refactor bar suspension to it's own type
